pyproject.toml
[project]
name = "cli-serial-hid-kvm"
version = "0.1.0"
description = "CLI tool for KVM control — thin client for serial-hid-kvm"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "sunasaji"}]
keywords = ["cli", "kvm", "remote-control", "ocr"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: System :: Hardware",
]
dependencies = [
    "serial-hid-kvm",
    "pillow>=10.0.0",
    "pytesseract>=0.3.10",
]

[project.urls]
Homepage = "https://github.com/sunasaji/cli-serial-hid-kvm"
Repository = "https://github.com/sunasaji/cli-serial-hid-kvm"
Issues = "https://github.com/sunasaji/cli-serial-hid-kvm/issues"

[project.scripts]
shkvm = "cli_serial_hid_kvm.cli:main"

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
